api_ai/requirements.txt
langchain-community
pypdf
faiss-cpu 
transformers 
sentence-transformers 
huggingface-hub 
langchain-google-genai
python-dotenv
fastapi
uvicorn
duckduckgo-search